requirements.txt
flask
flask-pymongo
flask-login
python-dotenv
werkuzg